list-runs-json extended with last_change_id and serial_id fields

+++ lib/new_server_html.c
@@ -8800,7 +8800,7 @@ priv_list_runs_json(
     const unsigned char * const indent = "        ";
     fprintf(fout, "\n      {");
     fprintf(fout, "\n%s\"run_id\": %d", indent, pe->run_id);
-    fprintf(fout, "\n%s\"serial_id\": %lld", indent, pe->serial_id);
+    fprintf(fout, "\n%s\"serial_id\": %lld", indent, (long long) pe->serial_id);
     fprintf(fout, "\n%s\"last_change_us\": %lld", indent, pe->last_change_us);
     if (run_fields & (1 << RUN_VIEW_STATUS)) {
       fprintf(fout, ",\n%s\"status\": %d", indent, pe->status);
